idk anymore
I won't add Creation Crystals here. The owner thing is still a bit odd. The ID is consistently unique between copies of the games, but clearly isn't entirely random. The last block of the 0x30 2011 remaining data bytes is also something to do with the owners and the way it is used. Sometimes modifying the owner ID doesn't seem to bring up the change ownership button if the aforementioned block sees no changes