gearbits
gearbits have some weird properties. if the value on the toy is also ever more than 33000, the game will cap it at 33000 and afterward often refuse to change the in game value even after being changed/lowered on the toy manually.

also idk why it's capped to the level 10 experience value, but it does correspond in offset to the 2013 experience area for regular tags, so it could be some sort of artefact of being built upon the regular core experience.